Implement changes CR
