+++ src/stringProcessing/parseSynonyms.js
@@ -4,7 +4,7 @@ const stripSpaces = require( "../stringProcessing/stripSpaces.js" );
 const removePunctuationExceptQuotes = require( "../stringProcessing/removePunctuationExceptQuotes.js" );
 
 /**
- * Matches strings from an array against a given text.
+ * Parses synonyms from a comma-separated string into an array.
  *
  * @param {String} synonyms The text to match
  *
